requirements.txt
openai==1.11.1
python-dotenv==0.10.1
fastapi
requests
beautifulsoup